requirements-secure.txt
# Core FastAPI dependencies
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0

# PDF processing
PyMuPDF==1.23.8
fpdf==1.7.2

# LLM and embeddings
openai==1.3.5
tiktoken==0.5.1
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
faiss-cpu==1.7.4

# Configuration
python-dotenv==1.0.0

# Security enhancements
python-magic==0.4.27  # For MIME type validation
slowapi==0.1.9  # For rate limiting (optional)

# Production server (optional)
gunicorn==21.2.0  # For production deployment

